pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-data-agent"
version = "0.1.0"
description = "Cockpit AI utterance synthesis powered by Hermes Agent skill system and self-evolution"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "ORBIT Team"},
]
keywords = ["hermes-agent", "cockpit", "nlu", "data-synthesis", "self-evolution"]
dependencies = [
    # ── Core: Hermes Agent ecosystem ──
    "hermes-agent @ git+https://github.com/NousResearch/hermes-agent.git",
    "hermes-agent-self-evolution @ git+https://github.com/NousResearch/hermes-agent-self-evolution.git",
    # ── Data processing ──
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    # ── CLI ──
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
hermes-data = "scripts.cli:main"

[tool.setuptools.packages.find]
include = ["tools*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
